Eliminate Meteor dependency in setting up Presenter In Pod
