+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/api/InMessages.scala
@@ -114,6 +114,12 @@ case class EjectUserFromBreakoutInternalMsg(parentId: String, breakoutId: String
 case class CapturePresentationReqInternalMsg(userId: String, parentMeetingId: String, filename: String, allPages: Boolean = true) extends InMessage
 
 /**
+ * Sent to the same meeting to force a new presenter to the Pod
+ * @param presenterId
+ */
+case class SetPresenterInDefaultPodInternalMsg(presenterId: String) extends InMessage
+
+/**
  * Sent by breakout room to parent meeting to obtain padId
  * @param breakoutId
  * @param filename
