+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/users/UsersApp.scala
@@ -2,12 +2,14 @@ package org.bigbluebutton.core.apps.users
 
 import org.apache.pekko.actor.ActorContext
 import org.apache.pekko.event.Logging
+import org.bigbluebutton.Boot.eventBus
 import org.bigbluebutton.common2.msgs._
+import org.bigbluebutton.core.api.{SetPresenterInDefaultPodInternalMsg}
 import org.bigbluebutton.core.apps.ExternalVideoModel
-import org.bigbluebutton.core.bus.InternalEventBus
+import org.bigbluebutton.core.bus.{BigBlueButtonEvent, InternalEventBus}
 import org.bigbluebutton.core.models._
 import org.bigbluebutton.core.running.{LiveMeeting, OutMsgRouter}
-import org.bigbluebutton.core2.message.senders.{MsgBuilder, Sender}
+import org.bigbluebutton.core2.message.senders.{MsgBuilder}
 import org.bigbluebutton.core.apps.screenshare.ScreenshareApp2x
 import org.bigbluebutton.core.db.UserStateDAO
 
@@ -67,14 +69,18 @@ object UsersApp {
       moderator <- Users2x.findModerator(liveMeeting.users2x)
       newPresenter <- Users2x.makePresenter(liveMeeting.users2x, moderator.intId)
     } yield {
-      // println(s"automaticallyAssignPresenter: moderator=${moderator} newPresenter=${newPresenter.intId}");
       sendPresenterAssigned(outGW, meetingId, newPresenter.intId, newPresenter.name, newPresenter.intId)
+      sendPresenterInPodReq(meetingId, newPresenter.intId)
     }
   }
 
   def sendPresenterAssigned(outGW: OutMsgRouter, meetingId: String, intId: String, name: String, assignedBy: String): Unit = {
     def event = MsgBuilder.buildPresenterAssignedEvtMsg(meetingId, intId, name, assignedBy)
     outGW.send(event)
+  }
+
+  def sendPresenterInPodReq(meetingId: String, newPresenterIntId: String): Unit = {
+    eventBus.publish(BigBlueButtonEvent(meetingId, SetPresenterInDefaultPodInternalMsg(newPresenterIntId)))
   }
 
   def sendUserLeftMeetingToAllClients(outGW: OutMsgRouter, meetingId: String,
