+++ bigbluebutton-html5/imports/api/users/server/handlers/presenterAssigned.js
@@ -1,21 +1,11 @@
 import Users from '/imports/api/users';
 import PresentationPods from '/imports/api/presentation-pods';
 import changePresenter from '/imports/api/users/server/modifiers/changePresenter';
-import RedisPubSub from '/imports/startup/server/redis';
+import Logger from '/imports/startup/server/logger';
 
 function setPresenterInPodReqMsg(credentials) { // TODO-- switch to meetingId, etc
-  const REDIS_CONFIG = Meteor.settings.private.redis;
-  const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-  const EVENT_NAME = 'SetPresenterInPodReqMsg';
-
-  const { meetingId, requesterUserId, presenterId } = credentials;
-
-  const payload = {
-    podId: 'DEFAULT_PRESENTATION_POD',
-    nextPresenterId: presenterId,
-  };
-
-  RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
+  // TODO It will be removed soon
+  Logger.debug(credentials);
 }
 
 export default async function handlePresenterAssigned({ body }, meetingId) {
